+++ samples/03_advances/07_lod/src/model_mesh.cpp
@@ -18,7 +18,8 @@ model_mesh::model_mesh(
 	_show_only_lod(false),
 	global_visiblity(true),
 	c_model(pContentPipelineModel),
-	_selected_lod_index(0)
+	_selected_lod_index(0),
+	_is_sky(false)
 	//++++++++++++++++++++++++++++++++++++++++++++++++++++
 	//++++++++++++++++++++++++++++++++++++++++++++++++++++
 {
@@ -1460,6 +1461,17 @@ W_RESULT model_mesh::_create_shader_modules(
 		V(W_FAILED, "loading fragment shader uniform 2 for model: " + this->model_name, _trace_info, 3);
 		return W_FAILED;
 	}
+	//draw sky without applying light and mipmaps
+	if (this->_is_sky)
+	{
+		this->_u2.data.cmds = 2;
+		auto _hr = this->_u2.update();
+		if (_hr == W_FAILED)
+		{
+			V(W_FAILED, "updating uniform u2(cmds) for sky: " + this->model_name, _trace_info, 3);
+		}
+	}
+
 	_shader_param.index = 3;
 	_shader_param.type = w_shader_binding_type::UNIFORM;
 	_shader_param.stage = w_shader_stage_flag_bits::FRAGMENT_SHADER;